refactor -> move CreateRoomForm to box package
